lets move the new tenant kube resource into a separate design file
